+++ src/services/Categories.php
@@ -16,9 +16,6 @@ use craft\elements\Entry;
 use craft\errors\FieldNotFoundException;
 use Faker\Factory;
 use Faker\Provider\Person;
-use studioespresso\seeder\records\SeederAssetRecord;
-use studioespresso\seeder\records\SeederCategoryRecord;
-use studioespresso\seeder\records\SeederEntryRecord;
 use studioespresso\seeder\Seeder;
 
 use Craft;
@@ -61,25 +58,19 @@ class Categories extends Component {
         }
 		$faker = Factory::create();
 
-		
+        $fields = Craft::$app->fields->getFieldsByElementType('craft\elements\Category');
+
 		for ( $x = 1; $x <= $count; $x ++ ) {
 			$category      = new Category( [
 				'groupId' => (int) $categoryGroup->id,
 				'title'     => Seeder::$plugin->fields->Title(20),
 			] );
-
+            Craft::$app->getElements()->saveElement( $category );
+			Seeder::$plugin->seeder->populateFields($fields, $category);
 			Craft::$app->getElements()->saveElement( $category );
-			$this->saveSeededCategory($category);
+			Seeder::$plugin->seeder->saveSeededCategory($category);
 		}
 
 	}
 
-    /**
-     * @param Asset $asset
-     */
-    public function saveSeededCategory($category) {
-        $record = new SeederCategoryRecord();
-        $record->categoryUid = $category->uid;
-        $record->save();
-    }
 }